requirements.txt
bcrypt==4.0.1
blinker==1.6.2
cffi==1.15.1
click==8.1.5
cryptography==41.0.2
Flask==2.3.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.2
Flask-User==1.0.2.0
Flask-WTF==1.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
passlib==1.7.4
pip==23.2
psycopg2
pycparser==2.21
python-dotenv==0.19.2
setuptools==65.5.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
Werkzeug==2.3.6
WTForms==3.0.1